Refactor: Consolidate RAG and recommendation APIs under /coach
This commit refactors the API by grouping RAG document storage and course recommendation endpoints under a single `/coach` category.

Key changes:
- The functionality from `api/app/routes/vectorstores.py` (RAG document storage) and `api/app/routes/assist.py` (course recommendation trigger) has been moved into `api/app/routes/coach.py`.
- The old `assist.py` and `vectorstores.py` files have been deleted.
- The main router in `api/app/main.py` has been updated to remove the old routes and register the new consolidated `/coach` router.
- The new API endpoints are now:
  - `POST /coach/documents` (previously `POST /vectorstores`)
  - `GET /coach/documents/{id}` (previously `GET /vectorstores/{id}`)
  - `POST /coach/documents/{id}/index` (previously `POST /vectorstores/{id}/index`)
  - `POST /coach/recommendations` (previously `POST /assist`)

Note: Verification of the running service was attempted, but the `docker compose up` command took an exceptionally long time to build in the execution environment, preventing a full live test of the new endpoints.
